Installation script improvement
This commit changes the installation script so that it is possible to change easily
the names of downloaded third-party packages. This is achieved by introduction
of variables of the archive file names. This change is very convenient for future
updates of the packages.
Moreover a new system of download and extract places is introduced. The packages
are extracted into the PSOPT root directory by default. After that the directories
of extracted packages are renamed from a complex versioned-name to a simple name
(without the version number). The compressed package files are then  moved into a
"Downloads" folder (created for that purpose in PSOPT root directory). According
to the names of the packages in the "Downloads" folder it is possible to get the
package version in the future.
This new version of the install script prevents also Adol-c compilation to create
a folder "adolc_base" in user home directory (which is really annoying). Instead of
that this directory is created in Adol-c root directory.
Personally I compiled PSOPT many times and the overall installation script never
worked for me at the first attempt. There were always issues with compilation errors
and it took a lot of manual work to adjust all paths and compiler flags etc. to went
through the whole script.
Based on this experience I made another modification, which makes the repairing procedure
a bit more easier. After each compilation the user is asked, wheter some errors are
shown in console. The advantage is that the user has now the information about what
was just compiled and if there was a success or not. If there are some errors one
can stop the script and try to repair the compilation errors.
The next small change is that I commented gnuplot installation procedure and left it
install by apt-get. For me it always worked the version from official repos. But if
neccessary the comments can be reverted again.
Apart from all that I needed for successful compilation of the dmatrix library to
add "-fpermissive" compiler flag (my gcc version is 9.2.1). This worked with my Linux
distro, which is a Debian sid. This amendment might not be neccessary with other
distros and compiler versions.